# api/http/api.yaml
openapi: 3.0.0
info:
  title: User Base
  version: 1.0.0
  description: API for managing user profiles
servers:
  - url: http://localhost:8080
    description: Local
paths:
  /profile:
    post:
      tags:
        - Profile
      operationId: createProfile
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileInput'
      responses:
        200:
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateProfileOutput'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Profile
      operationId: updateProfile
      summary: Update an existing profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileInput'
      responses:
        204:
          description: Profile updated
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    CreateProfileInput:
      type: object
      required:
        - name
        - age
        - email
        - phone
      properties:
        name:
          type: string
          description: Full name of the user
          example: John Doe
          minLength: 3
          maxLength: 64
        age:
          type: integer
          description: Age of the user
          example: 30
          minimum: 18
          maximum: 120
        email:
          type: string
          description: Email address of the user
          example: john.doe@example.com
        phone:
          type: string
          description: Phone number of the user
          example: +1234567890
    CreateProfileOutput:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the created profile
          example: 550e8400-e29b-41d4-a716-446655440000
    UpdateProfileInput:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier of the profile
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          nullable: true
          description: Full name of the user
          example: John Doe
          minimum: 18
          maximum: 120
        age:
          type: integer
          nullable: true
          description: Age of the user
          example: 30
          minimum: 18
          maximum: 120
        email:
          type: string
          nullable: true
          description: Email address of the user
          example: john.doe@example.com
        phone:
          type: string
          nullable: true
          description: Phone number of the user
          example: +1234567890
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
      example:
        error:
          error: Profile not found